highrise export button

+++ trunk/leads/app/View/Groups/index.ctp
@@ -43,6 +43,11 @@
 			</table>
 			
 			<a href="/Groups/mailmanExport/<?php echo $group['Group']['id']; ?>" onclick="return confirm('NOTE: Leads that have no email addresses or have already been moved to Highrise will NOT be exported.');">Export this Group to Mailman</a>
+			<?php if(empty($lead['highrise_id'])) { ?>
+				<a href="#" onclick="exportToHighrise(<?php echo $lead['id']; ?>); return false;" id="highrise-export">Export to Highrise</a>
+			<?php } else { ?>
+				<a href="https://jloop.highrisehq.com/companies/<?php echo $lead['highrise_id']; ?>" target="_blank" id="highrise-export">View in Highrise</a>
+			<?php } ?>
 			<a onclick="return confirm('Are you sure you want to delete this entire group? You\'re crazy.');" href="/Groups/delete/<?php echo $group['Group']['id']; ?>">Delete Group</a>
 		</div>
 	<?php } ?>
